+++ includes/admin/views/html-notice-upgrade-warning.php
@@ -25,8 +25,8 @@ if ( ! defined( 'ABSPATH' ) ) {
 			<p><?php echo sprintf( __( 'Version %1$s2.0.0%2$s of %3$s is coming soon and will provide a new and improved API. Before it is released, I require your help to test it out and provide your feedback!', 'cart-rest-api-for-woocommerce' ), '<strong>', '</strong>', 'CoCart' ); ?></p>
 		</div>
 
-		<div class="cocart-send-feedback">
-			<?php printf( '<a href="%1$s" class="button button-primary cocart-feedback-button" target="_blank">%2$s</a>', esc_url( 'https://cocart.xyz/cocart-v2-preview/' ), esc_html__( 'Learn More', 'cart-rest-api-for-woocommerce' ) ); ?>
+		<div class="cocart-action">
+			<?php printf( '<a href="%1$s" class="button button-primary cocart-button" target="_blank">%2$s</a>', esc_url( 'https://cocart.xyz/cocart-v2-preview/' ), esc_html__( 'Learn More', 'cart-rest-api-for-woocommerce' ) ); ?>
 			<a href="<?php echo esc_url( add_query_arg( 'hide_cocart_upgrade_notice', 'true' ) ); ?>" class="no-thanks"><?php echo esc_html__( 'Ask me again in 2 weeks', 'cart-rest-api-for-woocommerce' ); ?></a>
 		</div>
 	</div>