+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog for CoCart
 
-## v2.0.0 - ?? June, 2019
+## v2.0.0 - ?? ??, 2019
 
 * NEW: REST API namespace. CoCart is now an individual API and is no longer nested with WooCommerce's core REST API.
 * NEW: Check to see if the cart is set before falling back to the cart in session if one exists.
 * NEW: Get a specific customers cart via their customer ID number. - See documentation for details.
 * NEW: Product title also returns besides just the product name when getting the cart.
+* NEW: Product price also returns when getting the cart.
 * Changed: Filter and Action Hook names in new API. - See documentation for details.
 * Improved: Complexity of functions for better performance and usage.
 * Tweaked: Added checking for items already in the cart.
@@ -13,10 +14,12 @@
 * Tweaked: Responses for adding, updating, removing and restoring items to return whole cart if requested.
 * Tweaked: Totals can now return once calculated if requested.
 * Tweaked: Totals now return from session and can be returned pre-formatted if requested. - See documentation for details.
+* Tweaked: New option to refresh cart totals once item has been added.
 * Dev: Added action hooks for getting cart, cart is cleared, item added, item removed and item restored.
 * Dev: Added filter to allow additional checks before the item is added to the cart.
 * Dev: Added filter to apply additional data to return when cart is returned.
 * Dev: Added filter to change the size of the thumbnail returned.
+* Dev: Added new option to return cart raw if requested.
 
 ## v1.2.3 - 7th June, 2019